+++ pyubx2/ubxtypes_get.py
@@ -13,7 +13,8 @@ NB: Repeating groups must be defined as a tuple thus
        a) an integer representing a fixed number of repeats e.g 32
        b) a string representing the name of a preceding attribute
           containing the number of repeats e.g. 'numCh'
-       c) 'None' for an indeterminate number of repeats
+       c) 'None' for a 'variable by size' repeating group
+          (only one such group is permitted per message type)
     - {dict} is the nested dictionary containing the repeating
       attributes
 
@@ -21,7 +22,7 @@ Created on 27 Sep 2020
 
 @author: semuadmin
 """
-# pylint: disable=unused-import, too-many-lines, line-too-long
+# pylint: disable=too-many-lines, line-too-long
 
 from pyubx2.ubxtypes_core import (
     C2,